# Do not use syncPool for chunks when querying