--- a/db.go
+++ b/db.go
@@ -99,6 +99,7 @@ type DB struct {
 	logger    log.Logger
 	metrics   *dbMetrics
 	opts      *Options
+	chunkPool chunkenc.Pool
 	compactor Compactor
 
 	// Mutex for that must be held when modifying the general block layout.
@@ -241,6 +242,7 @@ func Open(dir string, l log.Logger, r prometheus.Registerer, opts *Options) (db 
 		donec:       make(chan struct{}),
 		stopc:       make(chan struct{}),
 		autoCompact: true,
+		chunkPool:   chunkenc.NewPool(),
 	}
 	db.metrics = newDBMetrics(db, r)
 
@@ -256,7 +258,7 @@ func Open(dir string, l log.Logger, r prometheus.Registerer, opts *Options) (db 
 		db.lockf = lockf
 	}
 
-	db.compactor, err = NewLeveledCompactor(r, l, opts.BlockRanges, chunkenc.NewPool())
+	db.compactor, err = NewLeveledCompactor(r, l, opts.BlockRanges, db.chunkPool)
 	if err != nil {
 		return nil, errors.Wrap(err, "create leveled compactor")
 	}
@@ -537,7 +539,7 @@ func (db *DB) reload() (err error) {
 		// See if we already have the block in memory or open it otherwise.
 		b, ok := db.getBlock(meta.ULID)
 		if !ok {
-			b, err = OpenBlock(dir, nil)
+			b, err = OpenBlock(dir, db.chunkPool)
 			if err != nil {
 				return errors.Wrapf(err, "open block %s", dir)
 			}