--- a/querier.go
+++ b/querier.go
@@ -603,6 +603,10 @@ func (s *populatedChunkSeries) Next() bool {
 			continue
 		}
 
+		for _, chk := range s.chks {
+			s.chunks.Put(chk.Chunk)
+		}
+
 		s.lset = lset
 		s.chks = chks
 		s.intervals = dranges