The only time any chunks are currently put into the syncPool is during a
compaction. This means that during almost all normal operation a new
chunk is still being created, along with the extra overhead of going
through the syncPool Get process.

This keeps the pool as an optional argument so as to not change the behavior of the compactor.

Noticed while working on #236 

When using the benchmarks from #425 we save a couple percent:
```
benchmark                                                          old ns/op     new ns/op     delta
BenchmarkPersistedQueries/series=10,samplesPerSeries=1000-8        807042        786265        -2.57%
BenchmarkPersistedQueries/series=10,samplesPerSeries=10000-8       7577470       7408384       -2.23%
BenchmarkPersistedQueries/series=10,samplesPerSeries=100000-8      75991905      75216135      -1.02%
BenchmarkPersistedQueries/series=100,samplesPerSeries=1000-8       7922229       7720921       -2.54%
BenchmarkPersistedQueries/series=100,samplesPerSeries=10000-8      76134169      74943460      -1.56%
BenchmarkPersistedQueries/series=100,samplesPerSeries=100000-8     782505185     763625329     -2.41%
```
And if we remove the `it.Next()` loop (focus only on chunk creation, not decoding) then we get:
```
benchmark                                                          old ns/op     new ns/op     delta
BenchmarkPersistedQueries/series=10,samplesPerSeries=1000-8        77687         57715         -25.71%
BenchmarkPersistedQueries/series=10,samplesPerSeries=10000-8       381082        188439        -50.55%
BenchmarkPersistedQueries/series=10,samplesPerSeries=100000-8      3467434       1542752       -55.51%
BenchmarkPersistedQueries/series=100,samplesPerSeries=1000-8       678544        477437        -29.64%
BenchmarkPersistedQueries/series=100,samplesPerSeries=10000-8      3814757       1870530       -50.97%
BenchmarkPersistedQueries/series=100,samplesPerSeries=100000-8     41222896      21328618      -48.26%
```

Signed-off-by: Chris Marchbanks <csmarchbanks@gmail.com>