@krasi-georgiev 2018-12-05
started a benchmark to test these at high load
https://github.com/prometheus/prometheus/pull/4962