@krasi-georgiev 2018-12-10
@csmarchbanks trying to ping you on IRC to syn on this.