@csmarchbanks 2018-12-10
I am at KubeCon this week. I will check in on IRC when i get a chance.