@csmarchbanks 2018-12-11
Results of promql benchmarks:
```
benchmark                                                                                                             old ns/op      new ns/op      delta
BenchmarkRangeQuery/expr=a_one,steps=1-4                                                                              20281          19789          -2.43%
BenchmarkRangeQuery/expr=a_one,steps=10-4                                                                             20909          19977          -4.46%
BenchmarkRangeQuery/expr=a_one,steps=100-4                                                                            27545          27506          -0.14%
BenchmarkRangeQuery/expr=a_one,steps=1000-4                                                                           91902          93073          +1.27%
BenchmarkRangeQuery/expr=a_ten,steps=1-4                                                                              82594          85832          +3.92%
BenchmarkRangeQuery/expr=a_ten,steps=10-4                                                                             85336          85363          +0.03%
BenchmarkRangeQuery/expr=a_ten,steps=100-4                                                                            156013         156167         +0.10%
BenchmarkRangeQuery/expr=a_ten,steps=1000-4                                                                           750001         737922         -1.61%
BenchmarkRangeQuery/expr=a_hundred,steps=1-4                                                                          722180         712345         -1.36%
BenchmarkRangeQuery/expr=a_hundred,steps=10-4                                                                         726530         720017         -0.90%
BenchmarkRangeQuery/expr=a_hundred,steps=100-4                                                                        1453659        1402699        -3.51%
BenchmarkRangeQuery/expr=a_hundred,steps=1000-4                                                                       7474817        7142547        -4.45%
BenchmarkRangeQuery/expr=rate(a_one[1m]),steps=1-4                                                                    26479          26554          +0.28%
BenchmarkRangeQuery/expr=rate(a_one[1m]),steps=10-4                                                                   27255          27052          -0.74%
BenchmarkRangeQuery/expr=rate(a_one[1m]),steps=100-4                                                                  44795          43905          -1.99%
BenchmarkRangeQuery/expr=rate(a_one[1m]),steps=1000-4                                                                 179728         177915         -1.01%
BenchmarkRangeQuery/expr=rate(a_ten[1m]),steps=1-4                                                                    97023          97742          +0.74%
BenchmarkRangeQuery/expr=rate(a_ten[1m]),steps=10-4                                                                   104369         105368         +0.96%
BenchmarkRangeQuery/expr=rate(a_ten[1m]),steps=100-4                                                                  272328         277905         +2.05%
BenchmarkRangeQuery/expr=rate(a_ten[1m]),steps=1000-4                                                                 1573604        1542411        -1.98%
BenchmarkRangeQuery/expr=rate(a_hundred[1m]),steps=1-4                                                                844297         828884         -1.83%
BenchmarkRangeQuery/expr=rate(a_hundred[1m]),steps=10-4                                                               909348         905053         -0.47%
BenchmarkRangeQuery/expr=rate(a_hundred[1m]),steps=100-4                                                              2437848        2364290        -3.02%
BenchmarkRangeQuery/expr=rate(a_hundred[1m]),steps=1000-4                                                             15436417       14736048       -4.54%
BenchmarkRangeQuery/expr=rate(a_one[1m]),steps=10000-4                                                                1627209        1580532        -2.87%
BenchmarkRangeQuery/expr=holt_winters(a_one[1d],_0.3,_0.3),steps=1-4                                                  984233         982967         -0.13%
BenchmarkRangeQuery/expr=holt_winters(a_one[1d],_0.3,_0.3),steps=10-4                                                 1493768        1500798        +0.47%
BenchmarkRangeQuery/expr=holt_winters(a_one[1d],_0.3,_0.3),steps=100-4                                                7055180        6539972        -7.30%
BenchmarkRangeQuery/expr=holt_winters(a_one[1d],_0.3,_0.3),steps=1000-4                                               62290127       56699422       -8.98%
BenchmarkRangeQuery/expr=holt_winters(a_ten[1d],_0.3,_0.3),steps=1-4                                                  7120073        6965394        -2.17%
BenchmarkRangeQuery/expr=holt_winters(a_ten[1d],_0.3,_0.3),steps=10-4                                                 12736240       11978701       -5.95%
BenchmarkRangeQuery/expr=holt_winters(a_ten[1d],_0.3,_0.3),steps=100-4                                                70068831       61893458       -11.67%
BenchmarkRangeQuery/expr=holt_winters(a_ten[1d],_0.3,_0.3),steps=1000-4                                               608929942      568155385      -6.70%
BenchmarkRangeQuery/expr=holt_winters(a_hundred[1d],_0.3,_0.3),steps=1-4                                              69020671       67410638       -2.33%
BenchmarkRangeQuery/expr=holt_winters(a_hundred[1d],_0.3,_0.3),steps=10-4                                             137341693      116824615      -14.94%
BenchmarkRangeQuery/expr=holt_winters(a_hundred[1d],_0.3,_0.3),steps=100-4                                            662860167      621487697      -6.24%
BenchmarkRangeQuery/expr=holt_winters(a_hundred[1d],_0.3,_0.3),steps=1000-4                                           6014163406     5653364109     -6.00%
BenchmarkRangeQuery/expr=changes(a_one[1d]),steps=1-4                                                                 818509         791261         -3.33%
BenchmarkRangeQuery/expr=changes(a_one[1d]),steps=10-4                                                                956269         972546         +1.70%
BenchmarkRangeQuery/expr=changes(a_one[1d]),steps=100-4                                                               2432807        2779940        +14.27%
BenchmarkRangeQuery/expr=changes(a_one[1d]),steps=1000-4                                                              17560289       21023055       +19.72%
BenchmarkRangeQuery/expr=changes(a_ten[1d]),steps=1-4                                                                 6115198        5981073        -2.19%
BenchmarkRangeQuery/expr=changes(a_ten[1d]),steps=10-4                                                                7599582        7873645        +3.61%
BenchmarkRangeQuery/expr=changes(a_ten[1d]),steps=100-4                                                               22265135       26054529       +17.02%
BenchmarkRangeQuery/expr=changes(a_ten[1d]),steps=1000-4                                                              169939013      208174582      +22.50%
BenchmarkRangeQuery/expr=changes(a_hundred[1d]),steps=1-4                                                             60943394       59257210       -2.77%
BenchmarkRangeQuery/expr=changes(a_hundred[1d]),steps=10-4                                                            76096495       77777998       +2.21%
BenchmarkRangeQuery/expr=changes(a_hundred[1d]),steps=100-4                                                           222767601      260934534      +17.13%
BenchmarkRangeQuery/expr=changes(a_hundred[1d]),steps=1000-4                                                          1708321160     2075568238     +21.50%
BenchmarkRangeQuery/expr=rate(a_one[1d]),steps=1-4                                                                    799779         771751         -3.50%
BenchmarkRangeQuery/expr=rate(a_one[1d]),steps=10-4                                                                   866910         865589         -0.15%
BenchmarkRangeQuery/expr=rate(a_one[1d]),steps=100-4                                                                  1761790        1799385        +2.13%
BenchmarkRangeQuery/expr=rate(a_one[1d]),steps=1000-4                                                                 10250762       10923610       +6.56%
BenchmarkRangeQuery/expr=rate(a_ten[1d]),steps=1-4                                                                    5991192        5833034        -2.64%
BenchmarkRangeQuery/expr=rate(a_ten[1d]),steps=10-4                                                                   6839318        6673881        -2.42%
BenchmarkRangeQuery/expr=rate(a_ten[1d]),steps=100-4                                                                  15349460       15865608       +3.36%
BenchmarkRangeQuery/expr=rate(a_ten[1d]),steps=1000-4                                                                 96806058       107417966      +10.96%
BenchmarkRangeQuery/expr=rate(a_hundred[1d]),steps=1-4                                                                60484520       57793673       -4.45%
BenchmarkRangeQuery/expr=rate(a_hundred[1d]),steps=10-4                                                               67420494       66817803       -0.89%
BenchmarkRangeQuery/expr=rate(a_hundred[1d]),steps=100-4                                                              153257860      158348889      +3.32%
BenchmarkRangeQuery/expr=rate(a_hundred[1d]),steps=1000-4                                                             973093143      1073376839     +10.31%
BenchmarkRangeQuery/expr=-a_one,steps=1-4                                                                             23492          22965          -2.24%
BenchmarkRangeQuery/expr=-a_one,steps=10-4                                                                            23172          23309          +0.59%
BenchmarkRangeQuery/expr=-a_one,steps=100-4                                                                           30212          29952          -0.86%
BenchmarkRangeQuery/expr=-a_one,steps=1000-4                                                                          97873          94745          -3.20%
BenchmarkRangeQuery/expr=-a_ten,steps=1-4                                                                             91237          88915          -2.55%
BenchmarkRangeQuery/expr=-a_ten,steps=10-4                                                                            93330          89744          -3.84%
BenchmarkRangeQuery/expr=-a_ten,steps=100-4                                                                           165839         163654         -1.32%
BenchmarkRangeQuery/expr=-a_ten,steps=1000-4                                                                          794324         735941         -7.35%
BenchmarkRangeQuery/expr=-a_hundred,steps=1-4                                                                         762791         738807         -3.14%
BenchmarkRangeQuery/expr=-a_hundred,steps=10-4                                                                        781680         760912         -2.66%
BenchmarkRangeQuery/expr=-a_hundred,steps=100-4                                                                       1543534        1470405        -4.74%
BenchmarkRangeQuery/expr=-a_hundred,steps=1000-4                                                                      7566888        7507028        -0.79%
BenchmarkRangeQuery/expr=a_one_-_b_one,steps=1-4                                                                      38834          37729          -2.85%
BenchmarkRangeQuery/expr=a_one_-_b_one,steps=10-4                                                                     44355          43238          -2.52%
BenchmarkRangeQuery/expr=a_one_-_b_one,steps=100-4                                                                    122334         117468         -3.98%
BenchmarkRangeQuery/expr=a_one_-_b_one,steps=1000-4                                                                   875395         817757         -6.58%
BenchmarkRangeQuery/expr=a_ten_-_b_ten,steps=1-4                                                                      201779         183602         -9.01%
BenchmarkRangeQuery/expr=a_ten_-_b_ten,steps=10-4                                                                     245855         234240         -4.72%
BenchmarkRangeQuery/expr=a_ten_-_b_ten,steps=100-4                                                                    904664         876541         -3.11%
BenchmarkRangeQuery/expr=a_ten_-_b_ten,steps=1000-4                                                                   7149383        6901341        -3.47%
BenchmarkRangeQuery/expr=a_hundred_-_b_hundred,steps=1-4                                                              1735501        1682093        -3.08%
BenchmarkRangeQuery/expr=a_hundred_-_b_hundred,steps=10-4                                                             2317435        2248205        -2.99%
BenchmarkRangeQuery/expr=a_hundred_-_b_hundred,steps=100-4                                                            9222119        9094203        -1.39%
BenchmarkRangeQuery/expr=a_hundred_-_b_hundred,steps=1000-4                                                           79840076       75631924       -5.27%
BenchmarkRangeQuery/expr=a_one_-_b_one,steps=10000-4                                                                  7814961        7674508        -1.80%
BenchmarkRangeQuery/expr=a_one_and_b_one{l=~'.*[0-4]$'},steps=1-4                                                     136394         117237         -14.05%
BenchmarkRangeQuery/expr=a_one_and_b_one{l=~'.*[0-4]$'},steps=10-4                                                    158431         120938         -23.67%
BenchmarkRangeQuery/expr=a_one_and_b_one{l=~'.*[0-4]$'},steps=100-4                                                   187403         155335         -17.11%
BenchmarkRangeQuery/expr=a_one_and_b_one{l=~'.*[0-4]$'},steps=1000-4                                                  556449         466422         -16.18%
BenchmarkRangeQuery/expr=a_ten_and_b_ten{l=~'.*[0-4]$'},steps=1-4                                                     300079         234865         -21.73%
BenchmarkRangeQuery/expr=a_ten_and_b_ten{l=~'.*[0-4]$'},steps=10-4                                                    351098         276558         -21.23%
BenchmarkRangeQuery/expr=a_ten_and_b_ten{l=~'.*[0-4]$'},steps=100-4                                                   671380         603499         -10.11%
BenchmarkRangeQuery/expr=a_ten_and_b_ten{l=~'.*[0-4]$'},steps=1000-4                                                  3640929        3613129        -0.76%
BenchmarkRangeQuery/expr=a_hundred_and_b_hundred{l=~'.*[0-4]$'},steps=1-4                                             1390400        1341526        -3.52%
BenchmarkRangeQuery/expr=a_hundred_and_b_hundred{l=~'.*[0-4]$'},steps=10-4                                            1770231        1632087        -7.80%
BenchmarkRangeQuery/expr=a_hundred_and_b_hundred{l=~'.*[0-4]$'},steps=100-4                                           5766364        5178335        -10.20%
BenchmarkRangeQuery/expr=a_hundred_and_b_hundred{l=~'.*[0-4]$'},steps=1000-4                                          44055655       40514024       -8.04%
BenchmarkRangeQuery/expr=a_one_or_b_one{l=~'.*[0-4]$'},steps=1-4                                                      153879         120855         -21.46%
BenchmarkRangeQuery/expr=a_one_or_b_one{l=~'.*[0-4]$'},steps=10-4                                                     155464         125061         -19.56%
BenchmarkRangeQuery/expr=a_one_or_b_one{l=~'.*[0-4]$'},steps=100-4                                                    234769         181216         -22.81%
BenchmarkRangeQuery/expr=a_one_or_b_one{l=~'.*[0-4]$'},steps=1000-4                                                   739278         641536         -13.22%
BenchmarkRangeQuery/expr=a_ten_or_b_ten{l=~'.*[0-4]$'},steps=1-4                                                      301155         242478         -19.48%
BenchmarkRangeQuery/expr=a_ten_or_b_ten{l=~'.*[0-4]$'},steps=10-4                                                     370152         286509         -22.60%
BenchmarkRangeQuery/expr=a_ten_or_b_ten{l=~'.*[0-4]$'},steps=100-4                                                    957670         741079         -22.62%
BenchmarkRangeQuery/expr=a_ten_or_b_ten{l=~'.*[0-4]$'},steps=1000-4                                                   5630522        4928022        -12.48%
BenchmarkRangeQuery/expr=a_hundred_or_b_hundred{l=~'.*[0-4]$'},steps=1-4                                              1722575        1352631        -21.48%
BenchmarkRangeQuery/expr=a_hundred_or_b_hundred{l=~'.*[0-4]$'},steps=10-4                                             2151703        1752790        -18.54%
BenchmarkRangeQuery/expr=a_hundred_or_b_hundred{l=~'.*[0-4]$'},steps=100-4                                            7413774        6565114        -11.45%
BenchmarkRangeQuery/expr=a_hundred_or_b_hundred{l=~'.*[0-4]$'},steps=1000-4                                           63843772       53478297       -16.24%
BenchmarkRangeQuery/expr=a_one_unless_b_one{l=~'.*[0-4]$'},steps=1-4                                                  198273         121315         -38.81%
BenchmarkRangeQuery/expr=a_one_unless_b_one{l=~'.*[0-4]$'},steps=10-4                                                 185580         127200         -31.46%
BenchmarkRangeQuery/expr=a_one_unless_b_one{l=~'.*[0-4]$'},steps=100-4                                                256993         175663         -31.65%
BenchmarkRangeQuery/expr=a_one_unless_b_one{l=~'.*[0-4]$'},steps=1000-4                                               752791         620880         -17.52%
BenchmarkRangeQuery/expr=a_ten_unless_b_ten{l=~'.*[0-4]$'},steps=1-4                                                  316572         239394         -24.38%
BenchmarkRangeQuery/expr=a_ten_unless_b_ten{l=~'.*[0-4]$'},steps=10-4                                                 343128         266319         -22.38%
BenchmarkRangeQuery/expr=a_ten_unless_b_ten{l=~'.*[0-4]$'},steps=100-4                                                722004         595300         -17.55%
BenchmarkRangeQuery/expr=a_ten_unless_b_ten{l=~'.*[0-4]$'},steps=1000-4                                               3838953        3586855        -6.57%
BenchmarkRangeQuery/expr=a_hundred_unless_b_hundred{l=~'.*[0-4]$'},steps=1-4                                          1521930        1308880        -14.00%
BenchmarkRangeQuery/expr=a_hundred_unless_b_hundred{l=~'.*[0-4]$'},steps=10-4                                         1831240        1581895        -13.62%
BenchmarkRangeQuery/expr=a_hundred_unless_b_hundred{l=~'.*[0-4]$'},steps=100-4                                        5765133        5212226        -9.59%
BenchmarkRangeQuery/expr=a_hundred_unless_b_hundred{l=~'.*[0-4]$'},steps=1000-4                                       47442414       40675768       -14.26%
BenchmarkRangeQuery/expr=abs(a_one),steps=1-4                                                                         39305          26148          -33.47%
BenchmarkRangeQuery/expr=abs(a_one),steps=10-4                                                                        38875          27943          -28.12%
BenchmarkRangeQuery/expr=abs(a_one),steps=100-4                                                                       79919          60590          -24.19%
BenchmarkRangeQuery/expr=abs(a_one),steps=1000-4                                                                      411899         347478         -15.64%
…